Move scope implementation to conceptImpl and add scope to type interface
