+++ grakn-core/src/main/java/ai/grakn/concept/EntityType.java
@@ -84,6 +84,22 @@ public interface EntityType extends Type{
      */
     Entity addEntity();
 
+    /**
+     * Classifies the type to a specific scope. This allows you to optionally categorise types.
+     *
+     * @param scope The category of this Type
+     * @return The Type itself.
+     */
+    EntityType scope(Instance scope);
+
+    /**
+     * Delete the scope specified.
+     *
+     * @param scope The Instances that is currently scoping this Type.
+     * @return The Type itself
+     */
+    EntityType deleteScope(Instance scope);
+
     //------------------------------------- Accessors ----------------------------------
     /**
      * Returns the supertype of this EntityType.
