+++ grakn-core/src/main/java/ai/grakn/concept/RoleType.java
@@ -79,6 +79,23 @@ public interface RoleType extends Type {
      */
     RoleType deletePlaysRole(RoleType roleType);
 
+
+    /**
+     * Classifies the type to a specific scope. This allows you to optionally categorise types.
+     *
+     * @param scope The category of this Type
+     * @return The Type itself.
+     */
+    RoleType scope(Instance scope);
+
+    /**
+     * Delete the scope specified.
+     *
+     * @param scope The Instances that is currently scoping this Type.
+     * @return The Type itself
+     */
+    RoleType deleteScope(Instance scope);
+
     //------------------------------------- Accessors ----------------------------------
     /**
      * Returns the supertype of this RoleType.
